// src/app/features/users/model/user.ts
export interface User {
    id: number;
    name: string;
    username: string;
    email: string;
    address: Address;
    phone: string;
    website: string;
    company: Company;
}

export interface Address {
    street: string;
    suite: string;
    city: string;
    zipcode: string;
    geo: GeoLocation;
}

export interface Company {
    name: string;
    catchPhrase: string;
    bs: string;
}

export interface GeoLocation {
    lat: string;
    lng: string;
}

export class UserDialogModel {
    constructor(public action: string, public actionClose: string, public data?: User) {}
}